Move sanitize transformer out of its special dir
